test when correct methods are called. Use the cheerio wrapper from render function.
